用户模块

+++ app/Http/Controllers/PostController.php
@@ -34,7 +34,9 @@ class PostController extends Controller
             'content' => 'required|string|min:10',
         ]);
 
-        $post = Post::create(request(['title', 'content']));
+        $user_id = \Auth::id();
+        $params = array_merge(compact('user_id'), \request(['title', 'content']));
+        $post = Post::create($params);
         return redirect("/posts");
     }
 
@@ -52,6 +54,9 @@ class PostController extends Controller
             'title' => 'required|string|max:100|min:5',
             'content' => 'required|string|min:10',
         ]);
+
+        $this->authorize('update', $post);
+
         $post->title = $request->get('title');
         $post->content = $request->get('content');
         $post->save();
@@ -62,7 +67,7 @@ class PostController extends Controller
     //删除逻辑
     public function delete(Post $post)
     {
-        // TODO:用户权限的验证
+        $this->authorize('delete', $post);
 
         $post->delete();
